(fix) Handle branches being out of sync
If master is out of sync (ie. not directly behind origin/master) then
it's probably because there are commits made locally directly on master
which haven't been pushed yet. In this case we want to fail the pull and
force the user to resolve this manually e.g. with a rebase or reset.

If gh-pages is out of sync then the user shouldn't care: gh-pages should
always only ever be master + the index.html file generated from master,
so we can freely reset this to point to the deployed commit.

Without these changes we'd end up creating merge commits, and probably
merging code we don't necessarily want.