(feature) Add redirect for deep links
When we started using Docsify back in February any links to a specific
section of the playbook that used a URL fragment (eg.
http://playbook.dxw.com/#sales) would result in a 404 error message.
This is because of the internal routing method that Docsify uses.

Unfortunately there's no official hook for a 404 in Docsify so we have
to do an AJAX call to see if the requested fragment is infact not a file
and probably just a section of the playbook.

Resolves: https://dxw.slack.com/archives/C025PM7N8/p1528286714000354