add Product management principles
this pull request adds the product community's principles to the playbook. It doesn't currently add an introductory 'we believe' statement - this can be added later in a separate pull request to be consistent with the principles for other professions, when the team has had a chance to think about what it should say.