(fix) Explicitly tell the search to look at the root path
When setting the 'auto' search mode with our current set up it will
break the search feature.

Since most of the searches will be on the main playbook I think this
will be ok. It means that search won't look for anything on the guides.

The alternative is to turn search off or replace Docsify.