Update to hosting and supporting services section
Following a review of the support section of the playbook, I've removed information that either needs to be included within internal documentation, for instance the ops docs or it needs to be re-worded to make it more accurate.

There is more to work on here - this update should reflect what should be removed, due to being out of date information or more internal focussed. Additions are a starting point to be iterated upon.