Rule out compressed hours for client-facing roles
This question came up recently, and after discussion with the directors
we concluded we are not happy to offer compressed hours for roles where
people are billed most of the time.

This change updates the policy in the playbook to reflect that decision.

The previous reason given in this document was due to the impact on
synchronous collaboration, but the reason was the problems it caused
with how we track and bill our time to clients from an admin overhead
perspective, and how we present that tracking and billing to clients
from a sales and client management perspective.

That complication doesn't exist for internal-facing roles, so those are
still open for discussion.