Create and add tone of voice guidelines
Created a new guide for tone of voice for the new brand. Apologies if this is a really long thing to review! Decided not to use the existing style guide as this info feels better as a new guide.

The original doc is here if you need it: https://docs.google.com/document/d/1Cn18K4Z8Ucq-UkzE0gMO6SsWPYAOmh-SisdENXnNamM/edit#