Fix syntax that was not intended as placeholder
Fixes the runtime warning
`Liquid Warning: Liquid syntax error (line 70): Expected end_of_string but found id in "{{A SPECIFIC PAGE}}" in guides/plugin-advisories.md`.

This did not come up during linting.

The text “a specific page” was not appearing, because it was being treated as a placeholder.

Use square brackets instead or curly, for consistency with the other example advisories, and capitalise similarly to the other advisories.

NB: The existing capitalisation goes against our style recommendations, but it is consistent throughout the plugin advisories guide.