Create and add tone of voice guidelines (#342)
* Create and add tone of voice guidelines

Created a new guide for tone of voice for the new brand. Apologies if this is a really long thing to review! Decided not to use the existing style guide as this info feels better as a new guide.

The original doc is here if you need it: https://docs.google.com/document/d/1Cn18K4Z8Ucq-UkzE0gMO6SsWPYAOmh-SisdENXnNamM/edit#

* Delete the existing style guide

We can merge relevant material from the existing style guide into the new Tone of voice guide.

* Added the new guide to the navigation

* Add a link to new tone of voice

Replace the existing style guidance with a reference to the new tone of voice guide.

* Reference new tone of voice guide

Replace a reference to the previous Playbook style guide with a reference to the new tone of voice

* Be consistent with Playbook

* Add guidance on names and technical terms

Add guidance on specific names and technical terms we use a lot so we can get rid of the old style guide.

Co-authored-by: John Waterworth <42607122+johnwaterworth@users.noreply.github.com>