Make sure sidebar links jump to the correct place
Before, we were using `scrollIntoView` to scroll to the target element. This worked,
but the headings were often obscured by the nav bar, as the main elements are
positioned absolutely on the page. This tweaks the code to get the location of the
element on the page and subtract the height of the navbar, so the element we've
scrolled to is visible.