Set focus correctly
When we're scrolling to the right part of the page when using the sidebar
navigation, the focus remains on the link, which can be confusing for screenreader
users. Because `window.scrollTo` doesn't have a callback function, we need to
use `setInterval` to poll to see if the page has been scrolled to roughly the right
place. If it has, we remove focus from the clicked link, and set focus to the target
heading element.

This has been tested on Voiceover on OSX and confirmed working.