Interim update to Working days (#632)
* Interim update to Working days

The Playbook still had some references to doing client work Monday through Thursday and to doing other things on Fridays. And to everyone on internal teams working 5 day weeks.

We're looking at policy and practice on this. But for now we should update this section to remove the out of date points.

* Apply automatic fixes

Co-authored-by: GitHub Actions <actions@github.com>