Move the CPI raise from April to October
We're shifting the CPI raise to October so that it comes after the end
of our financial year. That will let us use the financial performance of
the previous year to make a decision on whether or not we can afford a
CPI raise.

The details of this bit of the policy are likely to change after we
publish a pay policy, but this change should keep it broadly accurate
until then.