Don't autocommit fixes on PRs from forks
When a PR is opened from a fork (most commonly from Dependabot) any secrets
in the main git repo will not be available to the forked branch.

See https://github.blog/changelog/2021-02-19-github-actions-workflows-triggered-by-dependabot-prs-will-run-with-read-only-permissions/

Unfortunately this means this `test` Github action will always fail because it
autocommits fixes, and the secret required to autocommit fixes is not available
to the fork.

We don't think any PRs from Dependabot will require autocommitted linting fixes,
so it is safe to skip this action for PRs opened from forks.