Experimental: are we getting stuck on 'only modified'?
I'm wondering whether in the case of a PR which contains no
lintable changes, Vale hangs because it's waiting to hear
back from its 2 linters?

Perhaps this `onlyAnnotateModifiedLines` option is an
alternative if our intention is to limit feedback to the
content in the current PR?

i.e. everything will be linted on every PR but warnings and
errors will only be reported where they relate to changes in
the current PR.

Ref: https://github.com/dxw/playbook/pull/583/files