Prevent Vale from hanging
Currently, the Vale actions hangs when it has no files to lint, e.g.
Dependabot security PRs. It appears that if we just touch one of the
files after checkout (I chose README.md), that seems to work.

NB: It does mean that README.md need to pass Vale's linter, but this
seems like a reasonable compromise.