Vale is no longer run as part of CI
This action can create a huge amount of feedback that is very picky. For making changes to older files this is particularly challenging and introduces a lot of friction into making changes.

Firstly it means we have to go back through large documents and figure out how to rephrase sentences to use words Value _is_ happy about. Who knows how many iterations this would take, especially if you aren't set up to run local tests. Secondly it makes code review hard. The changes aren't scannable as every few lines there's a Vale notification complaining about something or another.

Using language that is then referred to as "weasley" is also a miserable experience from the authors point of view:

> 'regularly' is a weasel word!

Consider lowering the sensitivity of this check if we add it back in so it only notifies us of serious writing mishaps.