Vale tests all playbook files rather than just guides
I missed the directory within _includes and wrongly assumed all the content was in the guides directory in this earlier fix to CI [1]. This change runs vale on the index.md and _includes directory.

I couldn't see how to provide multiple paths to the same Vale configuration block so I decided to leave that as * and instead be specific about the directory when we invoke vale in the test file.

[1] https://github.com/dxw/playbook/pull/740