Fix headings being covered when using anchor links
Because we have a sticky header nav, we need to provide an offset when jumping to heading using an anchor link so that that the header doesn't cover the heading.

We had that offset in our JS, but not in our CSS. That meant that if you follow a link from outside the Playbook, the heading would be covered.

This change adds a `scroll-margin-top` property in the CSS to correct this, and changes the JS to use that value so that they are always the same.

The downside is that the offset is no longer linked to the size of the header, so if it got taller, this hardcoded number would also need to change.