Attempt to fix workflow (#850)
* Don't invoke husky in CI

The pre-commit hook runs the full linting suite, which is not what we want on CI.

The current issue this is attempting to solve is that the pre-commit hook invokes `vale` and crashes, because we don't install vale on CI. According to commit a43298962320b7ab70aa784c0d4807d89bb431b3, vale is not meant to be used on CI.

* First apply fixes, then build

When we use the repository's GITHUB_TOKEN to perform tasks on behalf of the GitHub Actions app, events triggered by the GITHUB_TOKEN will not create a new workflow run. This prevents us from accidentally creating recursive workflow runs.

However, this also leaves pull requests that have had automated fixes applied in an undefined state, where the test workflow isn't being run, so it isn't being reported, and the required status checks are left hanging.

The autocommit action will not create a commit unless the repo is "dirty" after running the linters in `fix` mode, so we don't need to use the failure of the previous action as a trigger for autocommitting. (That is the default behaviour, but I've included it to reassure reviewers.)

There could be better ways to check that we want those changes to be committed, but this isn't making anything worse from that point of view.