Rewrite the development workflow guide
Our existing guide was falling out of date with how we worked and was
very prescriptive and low level. This new guide zooms out a bit and
allows for more flexibility, leaning on providing guidance instead. We
lose some of the more step-by-step instructional content, which added a
lot of unnecessary bulk - new starters will have team members around to
show them how to do things.