Fold the sidebar on all screen sizes
There's something psycological about having looking at the menu be an
active choice to make that also makes it more useful. Open the menu to
find something in it, make your selection, and then close it - means
you're actually engaging with its presence and not having a second
distracting wall of headings floating on the page at all times.