Revert sidebar background to cream on large screens
In the midst of the Great Playbook Restructure of 2022, our dear benevolent
orchestrator F had split the large screen navigation sidebar in two: one column
dedicated to pages and another to in-page navigation. In order to visually
differentiate the two columns, the left was given the same blue background colour
as the top nav, while the right retained the cream of the original single column
sidebar design. In consultation with dxw brass, a decision was made to do away
with the in-page navigation altogether, but the remaining page navigation
remained blue, da ba dee da ba di

The original cream sidebar provided greater visual differentiation from the top
nav, and contributed to a more consistent quadrilateral colour block design,
whereas the shared blue background of the sidebar and top nav lead to a
six-sided L shape and two quadrilaterals, and a preponderance of blue, da ba dee
da ba di

Meanwhile in the Great Playbook Restructure of 2022, the navigation on mobile
moved up between the top nav and the main content, and become yellow, per the
square in which the D marque sits. This plays nice with the colours in the top
nav, and still provides good visual differentiation between the different
sections of the page, more so than would a thin bar of cream. This therefore
remains unchanged