Style search results and update excerpt
We've made the matching search term bold in search results. See the
example of this style below:

@see https://www.gov.uk/search/all?keywords=energy&order=relevance

The differentiation between bold and regular weight with our colour
palette is not as strong as in the example above. Therefore, we also
recoloured the search results to mid-blue. We don't think this looks
perfect, but we do think it's more legible

This also fetches the surrounding 200 characters to be displayed as an
excerpt. These are aligned with word boundaries, by moving outward until
a space is found, then trimming the excerpt to exclude the space

Unfortunately Markdown isn't being rendered here as this is an HTML file
(Jekyll's Markdown parsing is only applied to `.md` files) so we may
need to add our own Markdown parsing here, or find another way of using
Jekyll's built-in one.

Detruncate words in search result excerpts

The list style is removed both from search results and the messages
shown if there are no results

Co-authored-by: Ynda Jas <yndajas@gmail.com>