Italicise text matching query in search results
with surrounding 200 characters.

We'll probably want to highlight this in future to make it stand out,
but this is a first step.

Unfortunately Markdown isn't being rendered here as this is an HTML file
(Jekyll's Markdown parsing is only applied to `.md` files) so we may
need to add our own Markdown parsing here, or find another way of using
Jekyll's built-in one.