Restyle matching search results
In the search results, previously the matching term was bold and underlined,
which made it look like a link. Per accessibility advice, this keeps the font
weight but removes the text decoration styling. See the example of this style below

@see https://www.gov.uk/search/all?keywords=energy&order=relevance

In doing this, the matching words ceased to pop - the differentiation between
bold and regular weight with our colour palette is not as strong as in the
example above. Therefore, I also recoloured the search results to mid-blue. I
don't think this looks perfect, but I do think it's more legible