Move search and display calls out of loop
Previously the calls to search for results and display results were inside a
loop that iterated over the keys in `window.store`, which had one entry per page
in `site.pages`. The purpose of the loop was was to index all the pages with
Lunr (and format their content - a more recent addition), but it was also
performing the search and display after every addition to the index (i.e. once
per page). This mistake appears to have been copied over from the original guide
to implementing search in Jekyll with Lunr

This moves the search and display calls outside of the formatting and indexing
loop, avoiding wasteful client-side processing and making the page noticeably
snappier/more performant