FIX 404 errors on admin pages
Prior to this commit, when running netlify-cms locally, this bug
could not be reproduced. On production 404 errors were being
received from the GitHub API page, e.g.

https://api.github.com/repos/dxw/playbook/git/trees/main:product-management

returned a 404, whereas this URL:

https://api.github.com/repos/dxw/playbook/git/trees/main:src/product-management

returns 200 OK. So it seems that the GitHub back-end needs paths
relative to the repo root, so that it can correctly query the
GitHub API.

The documentation isn't very clear, but this page:

https://www.netlifycms.org/docs/collection-types#file-collections

says:

"""
Each file has its own list of fields and a unique filepath specified in
the file field (relative to the base of the repo).
"""

There isn't an equivalent statement for folder collections.

This commit changes the file paths, and also updates the
instructions for running the CMS features locally (where paths
can be relative to the content folder).