Highlight all search term matches in excerpt
Previously only the first match of the exact search query (in full) would be
highlighted. This changes the behaviour to highlight all matches of all search terms

It also adjusts the `matchIndex` to 100 if there are no matches in the content
(this is the case in results where the match is in the page title but not the
content). Adjusting the matchIndex to 100 ensures the excerpt in such cases is
~200 characters, rather than ~100 (previously if there was no match, the
`matchIndex` equalled `-1`)