Clean up search box outline
This removes the default outline of the search box, which had curved edges
(going against the Playbook's implicit design principles) and was partially
obscured by the search button

Instead, since the input and button are tightly coupled (visually and
conceptually), the entire search form is outlined when focusing within the
search box, with a cleaner design using one of our brand colours. This is the
same colour as used for underlining links, so the colour consistently signals
a location of (potential) action (clicking a link; searching the playbook)

The form outline has a short `ease-in-out` transition on the outline colour and
width, but this is skipped if the user prefers reduced motion

The default outline is retained on the submit button (when `:focus-visible`) in
order to avoid confusion when navigating the site by keyboard. When inside the
text input, the text cursor and form outline combine to indicate the focused
element, but the submit button doesn't have any analogue to a cursor to help
clarify which element within the form is focused, so maintaining the default
outline is more important