Clarify approach to code reviews
Previously the text on the process for merging a pull request referred to 2 reviewers including the author(s), which could be confusing. It was written in this way to accomodate a pair and mob situation- which is clarified int his update.

Reword second sentence slightly to improve readability, as there seems to be a missing word.