Update getExcerpt to prioritise excerpts containing full search phrase
Currently a search excerpt is generated by searching for the index of each separate term and selecting the first match for the first word (or the first match for the second if it is not found).

Here have added a search for the full term, so that an excerpt containing a full match will be prioritized over one containing just one word. The function uses two variables: information for the first (if any) full match, and a list of information on the first match for each word in the search query. Even if a full match is found, the list of individual term matchers is used for highlighting terms in the excerpt display.