Add babel and webpack
Configure basic webpack setup with multiple entry points to allow compilation of JavaScript with ES6 modules with imports.

This will allow us to install libraries through npm and import into the relevant file, rather than loading directly using a html script tag.

`npm start` script for serving the app locally currently uses a timer between webpack bundle and jeyll serve, as the app was often being served before the assets were bundled when using the `&` tag to run the tasks in parallel. This allows us to utilise live reload when making changes to both js and md files, but perhaps there is a better solution to be found here than using a timer.