Provide access to last updated date
According to the playbook[1]:

> The owner of each section of the Playbook should lead a review of the
  sections they own at least every quarter
>
> To help facilitate regular review, we should aim for Playbook content
  to include last updated and last reviewed dates.

The latter is not currently done, and the former is likely not done
consistently. Doing the latter will help us track how well we're
meeting the expectations we set out for reviewing playbook content on a
regular basis

@johnwaterworth raised an issue related to this: #1135

---

This gem allows us to get the date of the last git commit on the file
for a given page, or failing that when the file itself was modified.
This will feed the last updated information on each page

I would have liked to include a default time format in the config, but
this didn't work using the syntax in the gem's README[2]

[1]: https://playbook.dxw.com/contributing/managing-the-playbook/#review
[2]: https://github.com/gjtorikian/jekyll-last-modified-at#setting-up