Show last reviewed and updated date on frontend
This takes the last updated and last reviewed dates exposed in the
previous two commits and adds them to the bottom of each contentful page
(i.e. anything that has its own content beyond an index of pages in its
section). These are added at the end of the page after a horizontal rule

If the page has been updated since the last review, it will only show
this date. If a review has been recorded since the last update, it will
show both dates