Link to GitHub history of page
The previous commit added the last updated date to the bottom of each
contentful page. This adds a link to the respective GitHub history page,
for more detailed information (e.g. to check that the last update was
meaningful and not just a dev fixing formatting across multiple files)

The last updated date could in some cases give a misleading sense of
security that the page is more up to date than its content, so this
provides a quick way to check