Fix a few redirects
This includes adding the redirect script to a couple of pages. These pages have no
content, but are kept so that if someone looks for information on sick leave
or pay in the leave or pay, pension and benefits sections respectively, a link
will appear where expected. They redirect to the relevant policies keep things
DRY