DRY up permalink frontmatter
Before this change, on all Markdown files except `index.md` and
`**/index.md`, we specified a permalink of `/:path/:basename/`. The same
effect - the permalink being the path plus the kebabified name of the
file with no extension, except on index pages, for which it's just the
path - can be achieved with one config line. This updates the config and
removes the repeated line in each Markdown file