Remove 'Contact dxw' from index pages
Index pages (i.e. pages which have children but not their own content)
include the same partial as the nav menu. This menu includes a 'Contact
dxw' link, which looks a little out of place in the body of the index
page. This makes it so that the 'Contact dxw' link will only appear in
the nav, not the list of child pages on index pages