Update Playbook licence from CC BY-NC 2.0 to CC BY-NC 4.0
The Playbook is licensed with a Creative Common Attribution-NonCommercial 2.0 UK: England & Wales licence. That licence has since been replaced by version 4.0, which is an international licence and not region specific.

Creative Commons recommends that we use the newer version of the licence because it is "more user-friendly and more internationally robust"