Upgrade and add direct dependency: `sass-embedded`
There are no issues running the server locally, but Netlify is
struggling to install `sass-embedded`, which is blocking our Ruby
upgrade. The accessibility manual is at 1.77.2, which installs fine on
Netlify. This upgrades the gem to 1.77.2 for the playbook

This was removed as a direct dependency in 033d3be when addressing a
local development compatibility issue, but evidently it's remained or
reappeared as an indirect dependency. This restores it as a direct
dependency (originally added in a8981e9), in line with the accessibility
manual, which should mean that Renovate keeps it up to date. Everything
continues to run fine locally