Fix contributing links
At present, when you're on the "Benefits" page, clicking on "Add new page"
will take you to a form for adding a new "Pay, pension, and benefits"
page, the parent of "Benefits". This might have made sense in a previous
iteration of the menu/layout, but at present it's confusing. This
updates things so that:
- the homepage only allows you to edit itself, not add a new page (no
  change)
- other pages with children allow you to edit the page, and clicking
  "Add new page" will take you to the CMS page for adding a child of
  that page (not a sibling, as is presently the case)
- other pages without children will only allow you to edit the page

We aim to keep the `src/admin/config.yml` up-to-date with pages that
have children, in order to allow creation of new children via the CMS,
but in cases where we haven't kept things up to date, bad CMS links
should simply redirect to the main CMS page