Move scss property to above nested property
Running lint on the project stylesheets give the following warning:

```
Deprecation Warning: Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.
```

This change moves our one declaration which breaks this rule, and does not affect the behaviour (contribution links should only display on large screens).