Update authenticated hrefs to exclude in html check
Some external links in the playbook are authenticated due to privacy level, or will be directed to an authentication screen for other reasons (e.g. LinkedIn). Updating the list to include more domains that currently fail the script for this reasons.

Replacing the dxw private repository links with a catch-all for the dxw GitHub organisation, as there quite a few links to private repositories and adding them all will be harder to maintain.