Add html proofer workflow
Currently it is tricky to run checks like this in the merge pipeline, as we do not want to slow down contributions from members of staff using the Netlify CMS workflow.

However, it would be useful to run the HTML proofer more regularly to catch redundant links in the playbook. I have started with a manual workflow_dispatch, but we could update this to a schedule (note a schedule will disable if there have been no repository updates in 60 days, or if the committer leaves the organisation).

It would also be nice to post the results to the playbook channel in slack, to encourage page owners to tackle redundant links.