Local Netlify server serves via https
In order to test configuration for HTTP headers, such
as strict-transport-policy we need to be able to run
a local server via https.

This commit adds that configuration in two parts:

* Netlify configuration for the local development server
* Local certificate files created with mkcert

In order to regenerate the certificates, install mkcert
and run:

    mkcert localhost

To run the local server with Netlify configuration, run:

    npm install netlify-cli -g

to install the local development tools (and 'nodenv rehash',
if you are using nodenv to manage Node applications) and then:

    netlify dev

to run the server. The application can then be found at:

    https://localhost:8888/

in your browser.