Fix pages not appearing in menu
These pages don't appear in the menu because of having apostrophes in
their filenames and/or titles (haven't quite gotten to the bottom of the
problem). I tried changing the character from `’` to `'` but that didn't
fix it. The easiest solution seems to be avoiding the contraction
altogether, which also brings these pages in line with the sibling "If
you are the secondary carer" page